% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/consolidate_barcode_tailfindr_and_nanopolish_data.R
\name{consolidate_barcode_tailfindr_and_nanopolish_data}
\alias{consolidate_barcode_tailfindr_and_nanopolish_data}
\title{Title}
\usage{
consolidate_barcode_tailfindr_and_nanopolish_data(decoded_barcodes,
  tailfindr_estimates, nanopolish_estimates)
}
\arguments{
\item{decoded_barcodes}{}

\item{tailfindr_estimates}{}

\item{nanopolish_estimates}{}
}
\description{
Title
}
