% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_krause_etal_rna002_data.R
\name{consolidate_krause_etal_rna002_data}
\alias{consolidate_krause_etal_rna002_data}
\title{Title}
\usage{
consolidate_krause_etal_rna002_data(decoded_barcodes, tailfindr_estimates,
  nanopolish_estimates, transcript_start_info)
}
\arguments{
\item{decoded_barcodes}{}

\item{tailfindr_estimates}{}

\item{nanopolish_estimates}{}

\item{transcript_start_info}{}
}
\description{
Title
}
