% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from GitHub release using piggyback.}
\usage{
download_data(release_version)
}
\arguments{
\item{release_version}{Which version of the data to download}
}
\description{
Download data from GitHub release using piggyback.
}
