% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download data from GitHub release using piggyback and return it as
a dataframe.}
\usage{
get_data(release_version, file_name)
}
\arguments{
\item{release_version}{Which version of the data to download}
}
\description{
Download data from GitHub release using piggyback and return it as
a dataframe.
}
