% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_krause_etal_dna_data.R
\name{consolidate_krause_etal_dna_data}
\alias{consolidate_krause_etal_dna_data}
\title{Title}
\usage{
consolidate_krause_etal_dna_data(decoded_barcodes_ff,
  tailfindr_estimates_ff, transcript_start_info_ff, tailfindr_estimates_st,
  transcript_start_info_st)
}
\arguments{
\item{decoded_barcodes_ff}{}

\item{tailfindr_estimates_ff}{}

\item{transcript_start_info_ff}{}

\item{tailfindr_estimates_st}{}

\item{transcript_start_info_st}{}
}
\description{
Title
}
