% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_workman_etal_data.R
\name{consolidate_workman_etal_data}
\alias{consolidate_workman_etal_data}
\title{Title}
\usage{
consolidate_workman_etal_data(tailfindr_estimates, nanopolish_estimates,
  transcript_start_info)
}
\arguments{
\item{tailfindr_estimates}{}

\item{nanopolish_estimates}{}

\item{transcript_start_info}{}
}
\description{
Title
}
